Carbon R6 class

The Carbon generator creates a new Carbon-object, which is the class containing all the app logic. The class is based on the R6 OO-system and is thus reference-based with methods and data attached to each object, in contrast to the more well known S3 and S4 systems.

Initialization

A new 'Carbon'-object is initialized using the new() method on the generator:

x <- carbon$new(code = clipr::read_clip())

Fields

Public Fields

Description of fields of the R6 object that can be set by the user can be found in the following page.

Methods

Interacting with Browser

$carbonate	Using RSelenium fetch the carbon image output

$browse	open $uri in the browser

Aesthetics

$set_template	set $template

$get_templates	get a list of possible templates

$set_window_control_theme	set $windows_control_theme

$get_windows_control_themes	get a list of possible window control themes

$set_font_family	set $font_family

$get_font_families	get a list of possible fonts

URI Building

$uri	construct valid carbon.js uri

$options	return all current carbon options

$encode	URL encode a string for the $uri

$tiny	Create tinyurl from $uri

$rtweet	Send tweets containing media generated from carbon

Webdriver Settings

$chromeOptions	construct a chromeOptions object

$chrome_start	start a chrome session

$chrome_stop	stop a chrome session

$start	start a RSelenium session

$stop	stop a RSelenium session

$stop_all	stop all active RSelenium sessions

$get_port	Get active port

$set_port	Set new port