Carbon object fields

Fields that can be changed in the Carbon object

Arguments

code

character, script to embbed, Default: NULL

palette

numeric, layout pallete, Default: c(r=171,g=184,b=195,a=1)

template

character, layout template, Default: 'seti'

add_window_control

boolean, add/remove window controls (circles on top left corner), Default: TRUE

window_control_theme

character, change window control themes, Default: 'none'

language

character, language, Default: 'r'

add_drop_shadow

boolean, add/remove dropshadow, Default: TRUE

drop_shadow_offset_y

numeric, shadow vertical offset (px), Default: 20

drop_shadow_blur_radius

numeric, shadow blur radius (px), Default: 68

auto_adjust_width

boolean, auto-audjust output width, Default: TRUE

padding_vertical

numeric, vertical padding (px), Default: 48

padding_horizontal

numeric, horizontal padding (px), Default: 32

add_line_number

boolean, add/remove line numbers, Default: FALSE

font_family

character, layout font family, Default: 'Hack'

font_size

numeric, font size (px), Default: 14

line_height_percent

numeric, relative space between lines (percent), Default: 133

square_image

boolean, output image is square, Default: FALSE

relative_export_size

numeric, image size in export relative to what is in the preview (1,2 or 4), Default: 1

add_watermark

boolean, add official carbon.js watermark, Default: FALSE

add_tinyurl

boolean, add tinyurl link to image directing to the carbon page that created the image, Default: FALSE

tinyurl_location

character, string with gravity value to place the annotate tinyurl address on image, Default: 'southeast'

tweet

boolean, tweet out the image using post_tweet

tweet_status

character, status to add to the tweet. If add_tinyurl is TRUE then a link to the url will be appended to the status automatically, Default: 'Created in R using the Carbonate (Package Icon)'

add_timestamp

boolean, add timestamp to output filename, Default: FALSE

carbons

list, storage of rendered carbon images, Default: NULL

download_path

character, path to download the images, Default: tempdir()

chrome_args

character, arguments passed to chrome browser, Default: c('--disable-gpu', '--window-size=1280,800')

chrome_pref

list, preferences passed to chrome browser, Default: list("profile.default_content_settings.popups" = 0L,"download.prompt_for_download" = FALSE,"download.default_directory" = tempdir())

rD

RSelenium Driver object

cDrv

Chrome Driver object from wdman

See also