Carbonate script lines to a carbon image

Main function of the package that invokes RSelenium to open a browser to the carbon.js uri, create an image and download the file.

.carbonate(self, private, file, path, code, rD)

Arguments

self
carbon self object

private
carbon private object

file
character, name of file to save image as

path
character, path to save the image

code
character, lines of script to make carbon image from

rD
RSelenium driver

Value

image object

Details

Script is passed to https://carbon.now.sh/ is downloaded to the tempdir() and appended to the list $carbons using RSelenium and Chrome.

See also

Examples

if(interactive()){
 x <- carbon$new('x <- 1')
 x$carbonate()
 }